# docs/openapi.yaml
openapi: 3.2.0
info:
  title: Bill Payment API
  description: |
    A comprehensive RESTful API for bill payment operations, designed for API automation practice.

    ## Overview
    This API provides endpoints to manage billers, bills, payments, payment methods, users, and file uploads.
    It supports various authentication methods, content types, and includes realistic business logic.

    ## Authentication
    The API supports multiple authentication methods. All endpoints except `/health` require authentication.

    ### API Key (Recommended for Testing)
    - **Header**: `X-API-Key: demo-api-key-123`
    - **Query Parameter**: `?api_key=demo-api-key-123`

    ### Bearer Token (JWT)
    ```
    Authorization: Bearer demo-jwt-token-456
    ```

    ### Basic Authentication
    ```
    Authorization: Basic ZGVtbzpwYXNzd29yZDEyMw==
    ```
    (Credentials: `demo:password123`)

    ### Cookie-Based Session (NEW)
    ```
    Cookie: session_id=demo-session-abc123
    ```
    Session cookies are validated for authentication. Use the demo session token above for testing.

    ### OAuth2 Client Credentials
    ```bash
    POST /oauth/token
    Content-Type: application/x-www-form-urlencoded

    grant_type=client_credentials&client_id=demo-client&client_secret=demo-secret-789
    ```

    ## Content Types
    The API supports multiple content types:

    ### JSON (Default)
    - Request: `Content-Type: application/json`
    - Response: `Accept: application/json`

...

...

...

...

...

    ```
...
servers:
...
paths:
...
  /v1/payments:
    get:
      tags: [Payments]
      summary: List payments
      description: Get a paginated list of payments
      operationId: listPayments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: user_id
          in: query
          schema:
            type: string
        - name: bill_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/PaymentStatus'
        - name: from_date
          in: query
          schema:
            type: string
            format: date-time
        - name: to_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentsListResponse'

    post:
      tags: [Payments]
      summary: Create/process a payment
      description: |
        Initiate a payment for a bill. The payment will be processed
        immediately and the status will be returned.

        **Note**: This is a simulated payment for demo purposes.
        ~90% of payments succeed, ~10% fail randomly.
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentInput'
            examples:
              upiPayment:
                summary: UPI Payment
                value:
                  billId: bill-001
                  amount: 599.00
                  paymentMethodType: upi
                  paymentMethodId: pm-upi-001
              cardPayment:
                summary: Credit Card Payment
                value:
                  billId: bill-001
                  amount: 599.00
                  paymentMethodType: credit_card
                  paymentMethodId: pm-card-001
      responses:
        '201':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              examples:
                success:
                  summary: Successful payment
                  value:
                    success: true
                    data:
                      id: pay-abc123
                      billId: bill-001
                      userId: user-demo-001
                      billerName: Airtel Postpaid
                      billNickname: My Mobile Bill
                      customerIdentifier: "9876543210"
                      status: completed
                      transactionId: TXN20251218ABC123
                      referenceNumber: REF123456789012
                      amount:
                        value: 599.00
                        currency: INR
                      convenienceFee: 0
                      totalAmount:
                        value: 599.00
                        currency: INR
                      paymentMethod:
                        id: pm-upi-001
                        type: upi
                        details: null
                      processedAt: "2025-12-18T10:30:00.000Z"
                      createdAt: "2025-12-18T10:30:00.000Z"
                failed:
                  summary: Failed payment
                  value:
                    success: true
                    data:
                      id: pay-xyz789
                      billId: bill-001
                      userId: user-demo-001
                      billerName: Airtel Postpaid
                      billNickname: My Mobile Bill
                      customerIdentifier: "9876543210"
                      status: failed
                      failure:
                        code: BANK_DECLINED
                        reason: Payment declined by bank
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          description: Bill or payment method not found
    additionalOperations:
      # OpenAPI 3.2 describes non-standard verbs here, keyed by the uppercase
      # method name. This is the native RFC 10008 form; the POST fallback at
      # /v1/payments/query documents the same operation for tooling that does
      # not yet render additionalOperations.
      QUERY:
        tags: [HTTP QUERY]
        summary: Search payments with a request body (native QUERY)
        description: |
          Native `QUERY /v1/payments` - RFC 10008. Safe, idempotent, and
          carries a JSON body. Identical in behaviour to
          `POST /v1/payments/query`; `meta.querySource` distinguishes them.
        operationId: queryPaymentsNative
        requestBody:
          required: false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryRequest'
              example:
                filter: { status: completed }
                sort: ["-createdAt"]
        responses:
          '200':
            description: Matching resources
            headers:
              Accept-Query:
                schema: { type: string }
              ETag:
                schema: { type: string }
              Link:
                schema: { type: string }
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/QueryResponse'
          '304':
            description: Not modified
          '400':
            description: Content-Type missing, or body inconsistent with it
          '415':
            description: Unsupported Content-Type
          '422':
            description: Well-formed body, but the query cannot be processed
  /v1/payments/stats:
    get:
      tags: [Payments]
      summary: Get payment statistics
      description: Get aggregated payment statistics
      operationId: getPaymentStats
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Payment statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatsResponse'
  /v1/payments/{id}:
    get:
      tags: [Payments]
      summary: Get payment by ID
      operationId: getPayment
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Payments]
      summary: Delete payment
      description: Delete a payment. Completed payments cannot be deleted.
      operationId: deletePayment
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Payment deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '400':
          description: Cannot delete completed payment
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/payments/{id}/refund:
    post:
      tags: [Payments]
      summary: Refund a payment
      description: Process a refund for a completed payment
      operationId: refundPayment
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  description: "Refund amount (default: full amount)"
                reason:
                  type: string
                  description: "Reason for refund"
              example:
                reason: "Customer requested refund"
      responses:
        '200':
          description: Refund processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'
        '400':
          description: Payment cannot be refunded (not completed)
  /v1/payments/{id}/cancel:
    post:
      tags: [Payments]
      summary: Cancel a payment
      description: Cancel an initiated or processing payment
      operationId: cancelPayment
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: Payment cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Payment cannot be cancelled

  # ============================================
  # PAYMENT METHODS
  # ============================================
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed in header
    ApiKeyQuery:
      type: apiKey
      in: query
      name: api_key
      description: API key passed as query parameter
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token
    BasicAuth:
      type: http
      scheme: basic
      description: Basic authentication (username:password)
    CookieAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: |
        Session cookie authentication. 
        Demo value: `session_id=demo-session-abc123`
    OAuth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes:
            read:all: Read access to all resources
            write:all: Write access to all resources
        password:
          tokenUrl: /oauth/token
          scopes:
            read:all: Read access
            write:all: Write access
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    PaymentId:
      name: id
      in: path
      required: true
      description: Payment ID
      schema:
        type: string
        example: pay-001
...
  schemas:
...
    PaymentStatus:
      type: string
      enum: [initiated, processing, completed, failed, cancelled, refunded]
    PaymentMethodType:
      type: string
      enum: [upi, credit_card, debit_card, net_banking, wallet]
...
    PaymentInput:
      type: object
      required:
        - billId
        - amount
      properties:
        billId:
          type: string
        amount:
          type: number
          minimum: 1
        currency:
          type: string
          default: INR
        paymentMethodId:
          type: string
        paymentMethodType:
          $ref: '#/components/schemas/PaymentMethodType'
    PaymentResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Payment'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PaymentsListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
security:
  - ApiKeyHeader: []
  - ApiKeyQuery: []
  - BearerAuth: []
  - BasicAuth: []
  - CookieAuth: []
  - OAuth2: []